fix: fecth only datat for program
